backend/src/personal_analytics_backend/database.py
# backend/src/personal_analytics/database.py
# backend/src/personal_analytics/database.py
from sqlmodel import SQLModel, create_engine, Session
import os
from dotenv import load_dotenv

load_dotenv()

# PostgreSQL connection
DATABASE_URL = os.getenv(
    "DATABASE_URL",
    "postgresql://analytics_user:personal_analytics_password@localhost:5432/personal_analytics"
)

engine = create_engine(DATABASE_URL)

def create_db_and_tables():
    SQLModel.metadata.create_all(engine)

def get_session():
    with Session(engine) as session:
        yield session