backend/pyproject.toml
[project]
name = "personal-analytics-backend"
version = "0.1.0"
description = "Backend for personal health analytics"
authors = [
    {name = "Tim Schäfer", email = "ts+code@rcmd.org"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.0", # PostgreSQL driver
    "python-dotenv>=1.0.0", # For environment variable management
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/personal_analytics_backend"]
